refactor blueprints (2/6): extract today blueprint
Six /today/* routes plus the _today_autoclear helper move to
blueprints/today.py. Registered in app.py alongside the tasks
blueprint.

Routes moved:
- GET /today/, /today (page)
- GET /today/count
- GET /today/data
- POST /today/star
- POST /today/complete

Internal helper:
- _today_autoclear — only callers are /today/data and /today/count,
  blueprint-internal.

`url_for('login')` in today_page() retained as-is — login is still at
app top-level. Will migrate to `url_for('cockpit.login')` in sub-phase 5.

Smoke-tested: /today/, /today/count, /today/data all 200; /publish
200 (cross-blueprint).
